#336: Setting 디렉토리 ViewController+Layout > View 네이밍 변경
