+++ Todoary-iOS/Todoary/Presentation/ViewControllers/Summary/CategoryViewController.swift
@@ -30,6 +30,7 @@ class CategoryViewController: BaseViewController, BaseViewControllerProtocol {
         
         super.viewDidLoad()
         
+        style()
         layout()
         initialize()
     }
