+++ Todoary-iOS/Todoary/Presentation/Views/Setting/AlarmSettingView.swift
@@ -1,5 +1,5 @@
 //
-//  AlarmSettingViewController+Layout.swift
+//  AlarmSettingView.swift
 //  Todoary
 //
 //  Created by 박지윤 on 2022/07/10.
