+++ Todoary-iOS/Todoary/Presentation/Views/Setting/PinNumberSettingView.swift
@@ -1,5 +1,5 @@
 //
-//  PinNumberSettingViewController+Layout.swift
+//  PinNumberSettingView.swift
 //  Todoary
 //
 //  Created by 송채영 on 2022/08/14.
