+++ Todoary-iOS/Todoary/Presentation/Views/Setting/SettingView.swift
@@ -1,5 +1,5 @@
 //
-//  SettingViewController + Layout.swift
+//  SettingView.swift
 //  Todoary
 //
 //  Created by 예리 on 2022/07/10.
